post-share: move component into blocks/
