+++ assets/stylesheets/_components.scss
@@ -45,6 +45,7 @@
 @import 'blocks/post-item/style';
 @import 'blocks/post-likes/style';
 @import 'blocks/post-relative-time/style';
+@import 'blocks/post-share/style';
 @import 'blocks/post-status/style';
 @import 'blocks/reader-author-link/style';
 @import 'blocks/reader-avatar/style';
@@ -324,7 +325,6 @@
 @import 'my-sites/post/post-image/style';
 @import 'my-sites/post-relative-time-status/style';
 @import 'my-sites/post-selector/style';
-@import 'my-sites/post-share/style';
 @import 'my-sites/post-type-list/style';
 @import 'my-sites/post-type-list/post-actions-ellipsis-menu/style';
 @import 'my-sites/post-type-list/post-type-post-author/style';
