+++ client/blocks/post-share/sharing-preview-modal.jsx
